backend/requirements.txt
Flask>=3.0.3
Flask-SQLAlchemy>=3.1.1
psycopg2-binary>=2.9.9  # PostgreSQL
SQLAlchemy>=2.0.34
argon2-cffi>=23.1.0
pyjwt==2.9.0
redis==5.1.1
flask-cors==5.0.0